Move code to package extension
