+++ Project.toml
@@ -12,7 +12,6 @@ DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Dictionaries = "85a47980-9c8c-11e8-2b9f-f7ca1fa99fb4"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
-EinExprs = "b1794770-133b-4de1-afb4-526377e9f4c5"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
 ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
@@ -31,6 +30,12 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
 TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
+
+[weakdeps]
+EinExprs = "b1794770-133b-4de1-afb4-526377e9f4c5"
+
+[extensions]
+ITensorNetworksEinExprsExt = "EinExprs"
 
 [compat]
 AbstractTrees = "0.4.4"
